added class user and edite other classes
