+++ src/Users/Professor.java
@@ -2,53 +2,20 @@ package Users;
 
 import java.sql.Date;
 
-public class Professor {
-	String name;
+public class Professor extends User {
+
 	String department;
 	Date hirringDate;
 	String PhDTopic;
-	String email;
-	String phoneNumber;
 	
-	
-	public Professor(String name, String department, Date hirringDate, String phDTopic, String email,
-			String phoneNumber) {
-		super();
-		this.name = name;
+	public Professor( String department, Date hirringDate, String phDTopic , String name, String email,String phoneNumber) {
+		super(name ,email,phoneNumber);
 		this.department = department;
 		this.hirringDate = hirringDate;
 		PhDTopic = phDTopic;
-		this.email = email;
-		this.phoneNumber = phoneNumber;
-	}
+		
+		}
 	
-	
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getName() {
-		return name;
-	}
-	public void setName(String name) {
-		this.name = name;
-	}
 	public String getDepartment() {
 		return department;
 	}
@@ -69,11 +36,10 @@ public class Professor {
 	}
 	
 	
-	public void notifyProfessor(String message) {
+	public void Update(String name) {
 		// do some stuff
-		
 	}
 	
-	//test2
+	
 	
 }
