+++ src/Users/Student.java
@@ -1,47 +1,11 @@
 package Users;
 
-public class Student {
-	String name;
+public class Student extends User{
 	int Id;
-	String email;
-	String phoneNumber;
-	
-	
-	
+
 	public Student(String name, int id, String email, String phoneNumber) {
-		super();
-		this.name = name;
-		Id = id;
-		this.email = email;
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
+		super(name,email,phoneNumber);
+		Id = id;	
 	}
 
 	public int getId() {
@@ -51,9 +15,8 @@ public class Student {
 	public void setId(int id) {
 		Id = id;
 	}
-	
-	
-	public void notifyStudent(String message) {
+
+	public void Update(String name) {
 		// do some stuff
 	}
