+++ src/Users/TA.java
@@ -1,47 +1,11 @@
 package Users;
 
-public class TA {
-	String name;
+public class TA extends User {
 	String department;
-	String email;
-	String phoneNumber;
-	
-	
 
 	public TA(String name, String department, String email, String phoneNumber) {
-		super();
-		this.name = name;
+		super(name,email,phoneNumber);
 		this.department = department;
-		this.email = email;
-		this.phoneNumber = phoneNumber;
-	}
-
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPhoneNumber() {
-		return phoneNumber;
-	}
-
-
-	public void setPhoneNumber(String phoneNumber) {
-		this.phoneNumber = phoneNumber;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
 	}
 
 	public String getDepartment() {
@@ -52,9 +16,7 @@ public class TA {
 		this.department = department;
 	}
 	
-	public void notifyTA(String message) {
+	public void Update(String name) {
 		// do some stuff
-		
 	}
-	
 }
